Cleaning Data.
In prepare_data I made a data frame for the consumption of electricity
This meant i also had to leave out many of the other columns except time
I also made a data frame for the avg temp in denmark for a given hour.
And converted time to same timezone.
merged into one large dataframe